+++ subsys/net/conn_mgr/events_handler.c
@@ -10,7 +10,6 @@ LOG_MODULE_DECLARE(conn_mgr, CONFIG_NET_CONNECTION_MANAGER_LOG_LEVEL);
 #include <errno.h>
 #include <zephyr/net/net_if.h>
 #include <zephyr/net/net_mgmt.h>
-
 #include "conn_mgr_private.h"
 
 extern uint16_t iface_states[CONN_MGR_IFACE_MAX];
