+++ tests/net/conn_mgr_conn/src/main.c
@@ -14,9 +14,9 @@
 
 #include <zephyr/ztest.h>
 #include <zephyr/net/net_if.h>
-#include <zephyr/net/conn_mgr_connectivity.h>
 #include <zephyr/net/conn_mgr.h>
-#include "conn_mgr_private.h"
+#include <zephyr/net/conn_mgr_connectivity.h>
+#include <zephyr/net/conn_mgr_connectivity_impl.h>
 #include "test_conn_impl.h"
 #include "test_ifaces.h"
